refactor: optimize HTTP request handling in Wasm environment
