+++ pkg/gateway/context/middleware.go
@@ -1,6 +1,9 @@
 package context
 
 import (
+	gocontext "context"
+	"net/http"
+
 	"github.com/labstack/echo/v4"
 	"github.com/medama-io/go-useragent"
 	"github.com/onsonr/sonr/crypto/mpc"
@@ -21,6 +24,14 @@ type GatewayContext struct {
 	turnstileSiteKey string
 }
 
+func GetGateway(c echo.Context) (*GatewayContext, error) {
+	cc, ok := c.(*GatewayContext)
+	if !ok {
+		return nil, echo.NewHTTPError(http.StatusInternalServerError, "Gateway Context not found")
+	}
+	return cc, nil
+}
+
 func UseGateway(env hway.Hway, ipc common.IPFS, db *hwayorm.Queries) echo.MiddlewareFunc {
 	return func(next echo.HandlerFunc) echo.HandlerFunc {
 		return func(c echo.Context) error {
@@ -37,4 +48,9 @@ func UseGateway(env hway.Hway, ipc common.IPFS, db *hwayorm.Queries) echo.Middle
 			return next(ctx)
 		}
 	}
+}
+
+func BG() gocontext.Context {
+	ctx := gocontext.Background()
+	return ctx
 }
