+++ pkg/gateway/context/profiles.go
@@ -99,27 +99,3 @@ func DeleteProfile(c echo.Context) error {
 	return nil
 }
 
-// ╭───────────────────────────────────────────────────────────╮
-// │            Create Profile (/register/profile)             │
-// ╰───────────────────────────────────────────────────────────╯
-
-// DefaultCreateProfileParams returns a default CreateProfileParams
-func DefaultCreateProfileParams() CreateProfileParams {
-	return CreateProfileParams{
-		TurnstileSiteKey: "",
-		FirstNumber:      0,
-		LastNumber:       0,
-	}
-}
-
-// CreateProfileParams represents the parameters for creating a profile
-type CreateProfileParams struct {
-	TurnstileSiteKey string
-	FirstNumber      int
-	LastNumber       int
-}
-
-// Sum returns the sum of the first and last number
-func (d CreateProfileParams) Sum() int {
-	return d.FirstNumber + d.LastNumber
-}
