+++ pkg/gateway/handlers/error_handler.go
@@ -1,0 +1,1 @@
+package handlers
