+++ pkg/gateway/handlers/index_handler.go
