+++ pkg/vault/context/wasm.go
@@ -1,7 +1,7 @@
 //go:build js && wasm
 // +build js,wasm
 
-package wasm
+package context
 
 import (
 	"encoding/base64"
@@ -10,7 +10,7 @@ import (
 	"github.com/labstack/echo/v4"
 )
 
-func ContextMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
+func WASMMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
 	return func(c echo.Context) error {
 		// Extract WASM context from headers
 		if wasmCtx := c.Request().Header.Get("X-Wasm-Context"); wasmCtx != "" {
