+++ pkg/vault/routes/routes.go
@@ -2,14 +2,12 @@
 // +build js,wasm
 
 // Package vault provides the routes for the Decentralized Web Node (...or Sonr Motr).
-package vault
+package routes
 
 import (
 	"github.com/labstack/echo/v4"
-
-	"github.com/onsonr/sonr/internal/config/motr"
 )
 
 // RegisterRoutes registers the Decentralized Web Node API routes.
-func RegisterRoutes(e *echo.Echo, config *motr.Config) {
+func Register(e *echo.Echo) {
 }
